Add attributes for added players and npcs (attributes may not be necessary for players. we could have an inspect command later. but this contains max health info which is good to have)

+++ src/Libraries/Shinobytes.Ravenfall.DataModels/NpcInstance.cs
@@ -5,12 +5,12 @@ namespace Shinobytes.Ravenfall.RavenNet.Models
     public class NpcInstance : Entity<NpcInstance>
     {
         private int health;
-        private long endurance;
+        private int endurance;
 
         public int NpcId { get; set; }
         public int TransformId { get; set; }
         public int Health { get => health; set => Set(ref health, value); }
-        public long Endurance { get => endurance; set => Set(ref endurance, value); }
+        public int Endurance { get => endurance; set => Set(ref endurance, value); }
         public NpcAlignment Alignment { get; set; }
         public int SessionId { get; set; }
     }
